+++ crypto_kem/ntrulpr761/avx2/crypto_verify_1167.c
@@ -27,10 +27,10 @@ int PQCLEAN_NTRULPR761_AVX2_crypto_verify_1167(const unsigned char *x, const uns
     diff |= _mm256_srli_epi32(diff, 16);
     diff |= _mm256_srli_epi64(diff, 32);
 
-    differentbits = _mm256_extract_epi8(diff, 0);
-    differentbits |= _mm256_extract_epi8(diff, 8);
-    differentbits |= _mm256_extract_epi8(diff, 16);
-    differentbits |= _mm256_extract_epi8(diff, 24);
+    differentbits = (unsigned int) _mm256_extract_epi8(diff, 0);
+    differentbits |= (unsigned int) _mm256_extract_epi8(diff, 8);
+    differentbits |= (unsigned int) _mm256_extract_epi8(diff, 16);
+    differentbits |= (unsigned int) _mm256_extract_epi8(diff, 24);
 
     return (int) (1 & ((differentbits - 1) >> 8)) - 1;
 }
