+++ crypto_kem/ntrulpr857/avx2/crypto_decode_857x1723.c
@@ -10,7 +10,7 @@ static inline int16 mullo(int16 x, int16 y) {
 }
 
 static inline int16 mulhi(int16 x, int16 y) {
-    return (x * (int32)y) >> 16;
+    return (int16) ((x * (int32)y) >> 16);
 }
 
 static inline __m256i add(__m256i x, __m256i y) {
@@ -79,7 +79,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
     a0 = mulhi(a0, 276) - mulhi(mullo(a0, -22580), 743); /* -374...374 */
     a0 += s[2 * i + 0]; /* -374...629 */
     a0 += (a0 >> 15) & 743; /* 0...742 */
-    a1 = (s[2 * i + 1] << 8) + s[2 * i] - a0;
+    a1 = (int16) ((s[2 * i + 1] << 8) + s[2 * i] - a0);
     a1 = mullo(a1, -3881);
 
     /* invalid inputs might need reduction mod 14044 */
@@ -100,7 +100,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
     a0 -= 436; /* -670..>37 */
     a0 += (a0 >> 15) & 436; /* -234...435 */
     a0 += (a0 >> 15) & 436; /* 0...435 */
-    a1 = (a2 << 6) + ((s[i] - a0) >> 2);
+    a1 = (int16) ((a2 << 6) + ((s[i] - a0) >> 2));
     a1 = mullo(a1, 2405);
 
     /* invalid inputs might need reduction mod 8246 */
@@ -117,7 +117,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
         a0 -= 436; /* -670..>37 */
         a0 += (a0 >> 15) & 436; /* -234...435 */
         a0 += (a0 >> 15) & 436; /* 0...435 */
-        a1 = (a2 << 6) + ((s[i] - a0) >> 2);
+        a1 = (int16) ((a2 << 6) + ((s[i] - a0) >> 2));
         a1 = mullo(a1, 2405);
 
         /* invalid inputs might need reduction mod 436 */
@@ -139,7 +139,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
         a0 -= 334; /* -501..>103 */
         a0 += (a0 >> 15) & 334; /* -167...333 */
         a0 += (a0 >> 15) & 334; /* 0...333 */
-        a1 = (a2 << 7) + ((s[i] - a0) >> 1);
+        a1 = (int16) ((a2 << 7) + ((s[i] - a0) >> 1));
         a1 = mullo(a1, -22761);
 
         /* invalid inputs might need reduction mod 334 */
@@ -160,7 +160,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
     a0 -= 292; /* -438..>125 */
     a0 += (a0 >> 15) & 292; /* -146...291 */
     a0 += (a0 >> 15) & 292; /* 0...291 */
-    a1 = (a2 << 6) + ((s[i] - a0) >> 2);
+    a1 = (int16) ((a2 << 6) + ((s[i] - a0) >> 2));
     a1 = mullo(a1, -3591);
 
     /* invalid inputs might need reduction mod 7229 */
@@ -177,7 +177,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
         a0 -= 292; /* -438..>125 */
         a0 += (a0 >> 15) & 292; /* -146...291 */
         a0 += (a0 >> 15) & 292; /* 0...291 */
-        a1 = (a2 << 6) + ((s[i] - a0) >> 2);
+        a1 = (int16) ((a2 << 6) + ((s[i] - a0) >> 2));
         a1 = mullo(a1, -3591);
 
         /* invalid inputs might need reduction mod 292 */
@@ -199,7 +199,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
         a0 -= 273; /* -410..>118 */
         a0 += (a0 >> 15) & 273; /* -137...272 */
         a0 += (a0 >> 15) & 273; /* 0...272 */
-        a1 = (a2 << 8) + s[i] - a0;
+        a1 = (int16) ((a2 << 8) + s[i] - a0);
         a1 = mullo(a1, 4081);
 
         /* invalid inputs might need reduction mod 273 */
@@ -218,7 +218,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
     a0 = mulhi(a0, -259) - mulhi(mullo(a0, -3971), 4225); /* -2178...2112 */
     a0 += s[1 * i + 0]; /* -2178...2367 */
     a0 += (a0 >> 15) & 4225; /* 0...4224 */
-    a1 = (a2 << 8) + s[i] - a0;
+    a1 = (int16) ((a2 << 8) + s[i] - a0);
     a1 = mullo(a1, 12161);
 
     /* invalid inputs might need reduction mod 438 */
@@ -272,7 +272,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
     a0 += s[1 * i + 0]; /* -33...287 */
     a0 = mulhi(a0, 16) - mulhi(mullo(a0, -1008), 65); /* -33...32 */
     a0 += (a0 >> 15) & 65; /* 0...64 */
-    a1 = (a2 << 8) + s[i] - a0;
+    a1 = (int16) ((a2 << 8) + s[i] - a0);
     a1 = mullo(a1, 4033);
 
     /* invalid inputs might need reduction mod 1723 */
@@ -391,7 +391,7 @@ void PQCLEAN_NTRULPR857_AVX2_crypto_decode_857x1723(void *v, const unsigned char
 
     /* R1 ------> R0: reconstruct mod 857*[1723] */
 
-    R0[856] = 3 * R1[428] - 2583;
+    R0[856] = (int16) (3 * R1[428] - 2583);
     s -= 428;
     i = 412;
     for (;;) {
